Move the verbal_autopsy app to va_explorer/va_data_management
