+++ va_explorer/users/tests/factories.py
@@ -2,7 +2,7 @@ import factory
 from django.contrib.auth import get_user_model, models
 from factory import DjangoModelFactory, Faker, Sequence
 from factory.fuzzy import FuzzyInteger
-from verbal_autopsy.models import Location
+from va_explorer.va_data_management.models import Location
 
 User = get_user_model()
