+++ va_explorer/va_data_management/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class VaDataManagementConfig(AppConfig):
+    name = 'va_explorer.va_data_management'
