--- verbal_autopsy/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class VerbalAutopsyConfig(AppConfig):
-    name = 'verbal_autopsy'
